Modularized the exporter code somewhat
Details:

* Moved logging support, exceptions, and the resource cache into their
  own modules.

* Moved the global variable VERBOSE_LEVEL into the new _logging.py
  module and it is now set via a new parameter of setup_logging().
  This was needed because setting global variables does not work
  across modules.
